feat(ecommerce-uikit): add DateFormat component that is integratied with falcon localization logic
