+++ packages/falcon-ecommerce-uikit/src/Locale/index.tsx
@@ -1,2 +1,3 @@
 export * from './LocaleContext';
 export * from './Price';
+export * from './DateFormat';
